:recycle: product 엔티티 rentPricePerDay 타입 BigDecimal 로 변경

+++ src/main/java/com/pedalgenie/pedalgenieback/domain/product/dto/response/GetProductQueryResponse.java
@@ -3,6 +3,8 @@ package com.pedalgenie.pedalgenieback.domain.product.dto.response;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.querydsl.core.annotations.QueryProjection;
 
+import java.math.BigDecimal;
+
 // 쿼리 결과 전용
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record GetProductQueryResponse(
@@ -11,7 +13,7 @@ public record GetProductQueryResponse(
         String name,
         Long shopId,
         String shopName,
-        Double rentPricePerDay,
+        BigDecimal rentPricePerDay,
         Boolean isRentable,
         Boolean isPurchasable,
         Boolean isDemoable,
